+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next
+
+- Added `numpy>=2.0` support.
+
 ## 0.5.5
 
 - Fixed a bug in the `cai_causal_graph.causal_graph.CausalGraph.add_edge` method that would allow the addition of an